Replace character_stats with round reports and align timestamp columns
Dropped the character_stats accumulator: round_reports now stores the frame's
per-round counters and its game mode, and the personal-stats matrix, rankings
and titles derive from it at query time instead of from an accumulator table.
Renamed the character and clan stamps (creation_time -> created_at,
last_login_time -> last_seen_at, notice_time -> notice_at), replaced
clans.emblem_wip with emblem_at, and dropped the instructor award latch in
favour of the announcement letter latching itself in the mail table.

Carries the accompanying reference-alignment work already in the tree:
accounts replacing users, the news/mail column renames, typed host-settings
columns, the timestamptz conversions and the Discord integration.

+++ src/Shared/Domain/Rankings/RankingModels.cs
@@ -11,7 +11,7 @@ namespace Mgo2Server.Shared.Domain.Rankings;
 /// <param name="Identifier">Character identifier, or clan identifier on the clan board.</param>
 /// <param name="Name">Name of the character or clan.</param>
 /// <param name="Value">Quantity the board sorts by.</param>
-public sealed record RankingEntry(int Rank, int Identifier, string Name, int Value);
+public sealed record RankingEntry(int Rank, long Identifier, string Name, int Value);
 
 /// <summary>
 /// A window of a board plus the size of the whole board, which is what drives the
@@ -28,4 +28,4 @@ public sealed record RankingPage(IReadOnlyList<RankingEntry> Entries, int Total)
 /// <param name="Identifier">Character identifier, or clan identifier on the clan board.</param>
 /// <param name="Name">Name of the character or clan.</param>
 /// <param name="Value">Quantity the board sorts by, in the client's units.</param>
-public readonly record struct RankingBoardRow(int Identifier, string Name, long Value);
+public readonly record struct RankingBoardRow(long Identifier, string Name, long Value);
